+++ Locales/enUS.lua
@@ -52,6 +52,7 @@ L["Minimap"] = true
 L["Misc Options"] = true
 L["MyChatAlert"] = true
 L["MyChatAlert Global Keywords"] = true
+L["MyChatAlert Globals"] = true
 L["Number delimiter"] = "."
 L["Number Header"] = "#."
 L["Number of alerts: %s"] = true